% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{highlightSATSNE}
\alias{highlightSATSNE}
\title{Highlights colData observation values in SATSNE embedding.}
\usage{
highlightSATSNE(
  out,
  A_data,
  B_data,
  colour_by = "dataset",
  highlight = NULL,
  alpha = 0.6,
  point_size = NULL,
  point_shape = 16,
  palette = NULL,
  legend_pos = "right",
  raster_points = TRUE,
  raster.dpi = 300
)
}
\description{
Highlights colData observation values in SATSNE embedding.
}
