% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSATSNE}
\alias{plotSATSNE}
\title{Plots SATSNE embedding.}
\usage{
plotSATSNE(
  out,
  A_data,
  B_data,
  colour_by = "dataset",
  point_size = NULL,
  point_shape = 16,
  palette = NULL,
  legend_pos = "right",
  raster_points = TRUE,
  raster.dpi = 300
)
}
\description{
Plots SATSNE embedding.
}
