% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepareSATSNE}
\alias{prepareSATSNE}
\title{Extracts dataframe of SATSNE coordinates with colData observation.}
\usage{
prepareSATSNE(
  x_sce,
  y_sce,
  shared_feats,
  assay_type = "logcounts",
  dimred = NULL,
  pca = is.null(dimred),
  shared_pca = TRUE,
  cosine_norm = TRUE,
  n_dimred = 50,
  ntop = 500,
  subset_row = NULL,
  scale = FALSE,
  xy_mnns = NULL,
  yx_mnns = NULL,
  nk1 = 10,
  nk2 = 10,
  coupled_period = 30,
  uncoupled_period = 20,
  mnn_weight = 1,
  normalise = TRUE,
  num_threads = 1,
  verbose = TRUE,
  return_logs = FALSE,
  ...
)
}
\description{
Extracts dataframe of SATSNE coordinates with colData observation.
}
