% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsatsne.R
\name{Rsatsne}
\alias{Rsatsne}
\title{Barnes-Hut implementation of t-Distributed Stochastic Neighbor Embedding}
\usage{
Rsatsne(X, ...)
}
\description{
Wrapper for the C++ implementation of Barnes-Hut t-Distributed Stochastic Neighbor Embedding. t-SNE is a method for constructing a low dimensional embedding of high-dimensional data, distances or similarities. Exact t-SNE can be computed by setting theta=0.0.
}
