% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSATSNE.R
\name{runSATSNE}
\alias{runSATSNE}
\title{TODO: Add BPPARAMS}
\usage{
runSATSNE(
  x_sce,
  y_sce,
  shared_feats,
  assay_type = "logcounts",
  dimred = NULL,
  pca = is.null(dimred),
  shared_pca = TRUE,
  cosine_norm = TRUE,
  n_dimred = 50,
  xy_mnns = NULL,
  yx_mnns = NULL,
  nk1 = 10,
  nk2 = 10,
  coupled_period = 30,
  uncoupled_period = 20,
  normalise = TRUE,
  num_threads = 1,
  verbose = TRUE,
  ...
)
}
\description{
TODO: Add BPPARAMS
}
